tank/app/requirements.txt
#psycorpg2
psycopg2-binary

